requirements.txt
nltk
pdfminer
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Uploads==0.2.1
idna==2.8
importlib-metadata==0.20
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.1
pdf2image==1.6.0
Pillow==6.1.0
pluggy==0.12.0
py==1.8.0
pycparser==2.19
pyparsing==2.4.2
pytest==5.1.2
redis==3.3.8
requests==2.22.0
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.5
zipp==0.6.0